Project1/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
distlib==0.3.9
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-nested-routers==0.94.1
filelock==3.16.1
idna==3.10
mysqlclient==2.2.6
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.6.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.28.0
